Start refactoring some stuff that should be in its own modules/files
