+++ src/game/grid/mod.rs
@@ -1,4 +1,7 @@
 pub mod grid_layout;
+pub mod movement;
+
+pub mod collision;
 
 use std::ops::{Add, Sub};
 
@@ -30,277 +33,6 @@ pub fn plugin(app: &mut App) {
     app.register_type::<(GridPosition, GridLayout)>();
 }
 
-/// Grid-based collision
-pub mod collision {
-    use bevy::prelude::*;
-
-    use crate::game::grid::movement::GridMovement;
-    use crate::game::grid::GridPosition;
-    use crate::game::spawn::level::LevelWalls;
-    use crate::AppSet;
-
-    pub fn plugin(app: &mut App) {
-        app.register_type::<GridCollider>();
-        // app.add_systems(Update, apply_collision_forces.in_set(AppSet::UpdateVirtualGrid));
-        app.add_systems(Update, simple_wall_collisions.in_set(AppSet::Update));
-    }
-
-    #[derive(Component, Reflect, Debug, Copy, Clone, PartialEq)]
-    #[reflect(Component)]
-    pub struct GridCollider {
-        radius: f32,
-    }
-
-    #[derive(Component, Reflect, Debug, Copy, Clone, PartialEq)]
-    #[reflect(Component)]
-    pub struct Immovable;
-
-    impl Default for GridCollider {
-        fn default() -> Self {
-            Self { radius: 1. }
-        }
-    }
-
-    /// https://deepnight.net/tutorial/a-simple-platformer-engine-part-1-basics/
-    pub fn simple_wall_collisions(
-        walls: Res<LevelWalls>,
-        mut query_player: Query<(&mut GridPosition, &mut GridMovement)>,
-    ) {
-        const COLLIDE_SUBGRID_DIST_POS: f32 = 0.2; // between 0..1, a fractional piece of the grid that we're allowed to move towards a wall while next to it
-        const COLLIDE_SUBGRID_DIST_NEG: f32 = -0.2; // between 0..1, a fractional piece of the grid that we're allowed to move towards a wall while next to it
-
-        let walls = &walls;
-        for (mut player, _movement) in query_player.iter_mut() {
-            let (cx, cy) = (player.coordinates.x as i32, player.coordinates.y as i32);
-            let (xr, yr) = (player.offset.x, player.offset.y);
-
-            if walls.collides(cx - 1, cy) && xr < COLLIDE_SUBGRID_DIST_NEG {
-                // println!("{cx} {xr} collided in x");
-                player.offset.x = COLLIDE_SUBGRID_DIST_NEG;
-            }
-            if walls.collides(cx, cy - 1) && yr < COLLIDE_SUBGRID_DIST_NEG {
-                // println!("{cy} {yr} collided in y");
-                player.offset.y = COLLIDE_SUBGRID_DIST_NEG;
-            }
-            if walls.collides(cx + 1, cy) && xr > COLLIDE_SUBGRID_DIST_POS {
-                // println!("{cx} {xr} collided in x");
-                player.offset.x = COLLIDE_SUBGRID_DIST_POS;
-            }
-            if walls.collides(cx, cy + 1) && yr > COLLIDE_SUBGRID_DIST_POS {
-                // println!("{cy} {yr} collided in y");
-                player.offset.y = COLLIDE_SUBGRID_DIST_POS;
-            }
-        }
-    }
-
-    /// https://deepnight.net/tutorial/a-simple-platformer-engine-part-2-collisions/
-    /// this didn't end up working as well as the simpler one above. Keeping for ref, we can always delete if we hate it
-    pub fn _apply_collision_forces(
-        walls: Res<LevelWalls>,
-        mut query_actors: Query<
-            (&GridPosition, &GridCollider, &mut GridMovement),
-            Without<Immovable>,
-        >,
-    ) {
-        const WALL_COLLIDER_RADIUS: f32 = 1.0; // assume walls all have a 1-grid-unit-radius circle around them
-        const REPEL_FORCE: f32 = 1.; // grid units per sec squared
-
-        for (actor_pos, actor_collider, mut actor_movement) in query_actors.iter_mut() {
-            // get any wall within 2 units of us as a fast distance check
-            let nearby_walls: Vec<Vec2> = walls
-                .wall_locations
-                .iter()
-                .map(|&w| GridPosition::new(w.x as f32, w.y as f32)._actual_coordinates())
-                .filter(|&w| {
-                    (w.x - actor_pos.coordinates.x).abs() <= 2.
-                        && (w.y - actor_pos.coordinates.y).abs() <= 2.
-                })
-                .collect();
-
-            for wall in nearby_walls.into_iter() {
-                let actor_pos = actor_pos._actual_coordinates();
-                let dx = actor_pos.x - wall.x;
-                let dy = actor_pos.y - wall.y;
-                let dist = ((dx * dx) + (dy * dy)).sqrt();
-
-                let max_collide_dist = actor_collider.radius + WALL_COLLIDER_RADIUS;
-                if dist >= max_collide_dist {
-                    continue;
-                }
-
-                println!(
-                    "Player at {},{} Colliding with wall at {},{} - dist {}, max_dist: {}",
-                    actor_pos.x, actor_pos.y, wall.x, wall.y, dist, max_collide_dist
-                );
-
-                let repel_power = (max_collide_dist - dist) / max_collide_dist;
-
-                let ang = dy.atan2(dx); // wtf?
-                let ddx = ang.cos() * repel_power * REPEL_FORCE;
-                let ddy = ang.sin() * repel_power * REPEL_FORCE;
-                actor_movement.velocity.x += ddx;
-                actor_movement.velocity.y += ddy;
-
-                println!("Applied collision force of {}, {} to player", ddx, ddy);
-            }
-        }
-    }
-}
-
-/// Grid-based movement
-pub mod movement {
-    use crate::game::grid::{GridLayout, GridPosition};
-    use crate::input::PlayerAction;
-    use crate::AppSet;
-    use bevy::prelude::*;
-    use leafwing_input_manager::prelude::ActionState;
-    use std::time::Duration;
-
-    pub fn plugin(app: &mut App) {
-        app.add_systems(Update, update_roll_timer.in_set(AppSet::TickTimers));
-        app.add_systems(
-            Update,
-            (respond_to_input, apply_movement)
-                .chain()
-                .in_set(AppSet::UpdateVirtualGrid),
-        );
-
-        app.add_systems(
-            Update,
-            set_real_position_based_on_grid.in_set(AppSet::UpdateWorld),
-        );
-
-        app.register_type::<GridMovement>();
-    }
-
-    #[derive(Component, Reflect, Debug, Copy, Clone, PartialEq)]
-    #[reflect(Component)]
-    pub struct GridMovement {
-        pub velocity: Vec2,
-        pub friction: f32,
-        pub acceleration_player_force: Vec2,
-        pub acceleration_external_force: Vec2,
-        pub acceleration_player_multiplier: f32,
-        pub is_rolling: bool,
-    }
-
-    #[derive(Component, Reflect, Debug, PartialEq)]
-    #[reflect(Component)]
-    pub struct Roll {
-        pub timer: Timer,
-        pub velocity_multiplier: f32,
-    }
-
-    impl Roll {
-        pub const ROLL_TIME: Duration = Duration::from_millis(300);
-    }
-
-    impl Default for Roll {
-        fn default() -> Self {
-            Self {
-                timer: Timer::from_seconds(0.5, TimerMode::Once),
-                velocity_multiplier: 3.0,
-            }
-        }
-    }
-
-    impl Default for GridMovement {
-        fn default() -> Self {
-            Self {
-                velocity: Vec2::ZERO,
-                friction: 0.85,
-                acceleration_player_force: Vec2::ZERO,
-                acceleration_external_force: Vec2::ZERO,
-                acceleration_player_multiplier: 66.,
-                is_rolling: false,
-            }
-        }
-    }
-
-    impl GridMovement {
-        pub fn current_force(&self) -> Vec2 {
-            self.acceleration_player_force + self.acceleration_external_force
-        }
-    }
-
-    pub fn respond_to_input(mut query: Query<(&ActionState<PlayerAction>, &mut GridMovement)>) {
-        for (action_state, mut movement) in query.iter_mut() {
-            let mut intent = Vec2::ZERO;
-
-            if action_state.pressed(&PlayerAction::MoveUp) {
-                intent.y += 1.0;
-            }
-            if action_state.pressed(&PlayerAction::MoveDown) {
-                intent.y -= 1.0;
-            }
-            if action_state.pressed(&PlayerAction::MoveLeft) {
-                intent.x -= 1.0;
-            }
-            if action_state.pressed(&PlayerAction::MoveRight) {
-                intent.x += 1.0;
-            }
-            // Normalize so that diagonal movement has the same speed as horizontal and vertical movement.
-            let intent = intent.normalize_or_zero();
-
-            movement.acceleration_player_force = intent * movement.acceleration_player_multiplier;
-
-            if !movement.is_rolling && action_state.pressed(&PlayerAction::Roll) {
-                movement.is_rolling = true;
-            }
-        }
-    }
-
-    pub fn apply_movement(
-        mut query: Query<(&mut GridPosition, &mut GridMovement, &mut Roll)>,
-        time: Res<Time>,
-    ) {
-        let dt = time.delta_seconds();
-        for (mut position, mut movement, roll) in query.iter_mut() {
-            let force = movement.current_force() * dt; // scale it by time
-
-            // apply forces and friction
-            let mut velocity = movement.velocity + force;
-            velocity *= movement.friction;
-            if velocity.length() < 0.01 {
-                velocity = Vec2::ZERO;
-            }
-            movement.velocity = velocity;
-
-            // move the player
-            let multiplier = if movement.is_rolling {
-                roll.velocity_multiplier
-            } else {
-                1.0
-            };
-            position.offset += movement.velocity * dt * multiplier;
-            position.fix_offset_overflow();
-        }
-    }
-
-    fn update_roll_timer(time: Res<Time>, mut query: Query<(&mut Roll, &mut GridMovement)>) {
-        let dt = time.delta_seconds();
-        for (mut roll, mut movement) in query.iter_mut() {
-            roll.timer.tick(Duration::from_secs_f32(dt));
-            if roll.timer.finished() {
-                movement.is_rolling = false;
-                roll.timer.reset();
-            }
-        }
-    }
-
-    /// Any entity that has a GridPosition and a Transform gets put in the world wherever its grid position says.
-    /// This does mean that Transform mutations get overwritten by grid position calculated ones.
-    pub fn set_real_position_based_on_grid(
-        mut query: Query<(&mut Transform, &GridPosition)>,
-        grid: Res<GridLayout>,
-    ) {
-        for (mut t, gp) in query.iter_mut() {
-            let pos = grid.grid_to_world(gp);
-            t.translation.x = pos.x;
-            t.translation.y = pos.y;
-        }
-    }
-}
 
 #[derive(Component)]
 pub struct GridSprite;
